refactor: replaced NothingFoundBg component and associated styles for a cleaner codebase
