--- client/src/components/core/NothingFoundBg/NothingFoundBg.module.css
@@ -1,60 +1,0 @@
-.root {
-  padding-top: 80px;
-  padding-bottom: 80px;
-}
-
-.fullScreen {
-  width: 100vw;
-  min-height: 100vh;
-  max-width: 100vw;
-  margin: 0;
-  padding: 0;
-}
-
-.inner {
-  position: relative;
-}
-
-.image {
-  position: absolute;
-  inset: 0;
-  opacity: 0.75;
-  color: light-dark(var(--mantine-color-gray-0), var(--mantine-color-dark-6));
-}
-
-.content {
-  padding-top: 220px;
-  position: relative;
-  z-index: 1;
-
-  @media (max-width: 576px) {
-    padding-top: 120px;
-  }
-}
-
-.title {
-  font-family: Outfit, var(--mantine-font-family);
-  text-align: center;
-  font-weight: 500;
-  font-size: 38px;
-
-  @media (max-width: 576px) {
-    font-size: 32px;
-  }
-}
-
-.description {
-  max-width: 540px;
-  margin: auto;
-  margin-top: var(--mantine-spacing-xl);
-  margin-bottom: calc(var(--mantine-spacing-xl) * 1.5);
-  text-align: center;
-  color: #868e96;
-  font-size: 18px;
-  font-weight: 400;
-  line-height: 1.5;
-  letter-spacing: 0.02em;
-  text-align: center;
-  text-wrap: balance;
-  text-wrap: balance;
-}
